handlers.py
from telegram.ext import CommandHandler, MessageHandler, filters, InlineQueryHandler
import image_manager
import pdf_converter
import ocr_utils
import drive_upload
import inline

def register(app):
    # Core commands
    app.add_handler(CommandHandler("start", start))
    app.add_handler(CommandHandler("help", help_cmd))
    app.add_handler(CommandHandler("convert", convert))
    app.add_handler(CommandHandler("list", list_images))
    app.add_handler(CommandHandler("remove", remove_image))
    app.add_handler(CommandHandler("move", move_image))
    app.add_handler(CommandHandler("options", set_options))
    app.add_handler(CommandHandler("title", set_title))
    app.add_handler(CommandHandler("watermark", set_watermark))
    app.add_handler(CommandHandler("addtext", add_text_page))
    app.add_handler(CommandHandler("password", set_pdf_password))
    app.add_handler(CommandHandler("upload", upload_pdf))
    app.add_handler(CommandHandler("ocr", ocr_images))
    app.add_handler(CommandHandler("stats", show_stats))
    app.add_handler(CommandHandler("language", set_language))
    app.add_handler(CommandHandler("startgroup", start_group))
    app.add_handler(CommandHandler("endsession", end_session))

    # Image/file handler
    app.add_handler(MessageHandler(filters.PHOTO | filters.Document.IMAGE, image_manager.handle_image))
    # Inline support
    app.add_handler(InlineQueryHandler(inline.inline_query))

# Implement each handler below (stubs for now)

async def start(update, context):
    await update.message.reply_text("Welcome! Send JPG/PNG images. Use /convert to get a PDF. Send /help for options.")

async def help_cmd(update, context):
    # List all commands and features
    await update.message.reply_text(
        "/start - Start the bot\n"
        "/convert - Convert sent images to PDF\n"
        "/list - List uploaded images\n"
        "/remove <n> - Remove image by index\n"
        "/move <from> <to> - Reorder images\n"
        "/options - Set PDF/page options\n"
        "/title <text> - Add title page\n"
        "/watermark <text> - Add watermark\n"
        "/addtext <text> - Add text page\n"
        "/password <pass> - Set PDF password\n"
        "/upload - Upload PDF to cloud\n"
        "/ocr - Extract text from images\n"
        "/stats - Your usage stats\n"
        "/language <code> - Change bot language\n"
        "/startgroup - Start group PDF session\n"
        "/endsession - End current session"
    )

async def convert(update, context):
    # Gather images, options, call pdf_converter, send PDF, clean up
    await pdf_converter.convert_images_to_pdf(update, context)

async def list_images(update, context):
    await image_manager.list_images(update, context)

async def remove_image(update, context):
    await image_manager.remove_image(update, context)

async def move_image(update, context):
    await image_manager.move_image(update, context)

async def set_options(update, context):
    await pdf_converter.set_options(update, context)

async def set_title(update, context):
    await pdf_converter.set_title(update, context)

async def set_watermark(update, context):
    await pdf_converter.set_watermark(update, context)

async def add_text_page(update, context):
    await pdf_converter.add_text_page(update, context)

async def set_pdf_password(update, context):
    await pdf_converter.set_pdf_password(update, context)

async def upload_pdf(update, context):
    await drive_upload.upload_pdf(update, context)

async def ocr_images(update, context):
    await ocr_utils.ocr_images(update, context)

async def show_stats(update, context):
    await image_manager.show_stats(update, context)

async def set_language(update, context):
    # Change user's preferred language
    pass

async def start_group(update, context):
    # Initiate group session
    pass

async def end_session(update, context):
    # Clean up session for user or group
    pass