pdf_converter.py
from fpdf import FPDF
from PIL import Image

# Store PDF options per user (page size, orientation, margins, compression, etc.)

async def convert_images_to_pdf(update, context):
    """Convert images to PDF using current session and options."""
    # Implement reordering, add title/watermark, handle options/password, etc.
    pass

async def set_options(update, context):
    """Set PDF options interactively (page size, orientation, etc.)."""
    pass

async def set_title(update, context):
    """Set a title page for the next PDF."""
    pass

async def set_watermark(update, context):
    """Set a watermark for all pages."""
    pass

async def add_text_page(update, context):
    """Add a page containing only text."""
    pass

async def set_pdf_password(update, context):
    """Set a password for the PDF output."""
    pass