requirements.txt
python-telegram-bot>=20.0
Pillow
fpdf2
pytesseract
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dropbox
msal
pycryptodome